requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
langchain-chroma
chromadb
tavily-python
google-generativeai
python-dotenv
pydantic